feat: get user relation
