--- src/user/dto/user-info.dto.ts
@@ -1,18 +1,0 @@
-import { Expose, Type } from 'class-transformer';
-
-export class UserInfoDto {
-  @Expose()
-  id: number;
-
-  @Expose()
-  email: string;
-
-  @Expose()
-  fullName: string;
-  
-  @Expose()
-  registrationDate: Date;
-
-  @Expose()
-  role: string;
-}
